draft: f3: adc: drop the need for extsel defines
Note: this is "compatible" with prior f3 code, but is completely different
behaviour to f4 for instance, so would be surprising and bad to make the change
as is.  it's just an idea.  This would however completley drop the idea of
tryign to make defines that cover all the complicated f3 cases.

Relates to github issue #518